+++ mybatis-rsql/src/main/java/com/talanlabs/mybatis/rsql/builder/impls/properties/FloatProperty.java
@@ -2,14 +2,14 @@ package com.talanlabs.mybatis.rsql.builder.impls.properties;
 
 import com.talanlabs.mybatis.rsql.builder.impls.AbstractField;
 
-public class IntegerProperty<Builder> extends AbstractNumberProperty<Builder, Integer> {
+public class FloatProperty<Builder> extends AbstractNumberProperty<Builder, Float> {
 
-    public IntegerProperty(String field, AbstractField<Builder> fieldBuilder) {
+    public FloatProperty(String field, AbstractField<Builder> fieldBuilder) {
         super(field, fieldBuilder);
     }
 
     @Override
-    protected String toString(Integer value) {
+    protected String toString(Float value) {
         return value != null ? value.toString() : "";
     }
 }
